+++ src/main/resources/META-INF/services/io.vertx.core.spi.cluster.ClusterManager
@@ -1,1 +1,1 @@
-io.vertx.spi.cluster.impl.zookeeper.ZookeeperClusterManager
+io.vertx.spi.cluster.zookeeper.ZookeeperClusterManager
