+++ src/test/java/io/vertx/test/core/ZKClusteredHATest.java
@@ -1,7 +1,7 @@
 package io.vertx.test.core;
 
 import io.vertx.core.spi.cluster.ClusterManager;
-import io.vertx.spi.cluster.impl.zookeeper.MockZKCluster;
+import io.vertx.spi.cluster.zookeeper.MockZKCluster;
 
 /**
  * Created by stream.Liu
